112 to 113 migration of lastBrowserPath + FileBrowser rotation fix
when migrating from 112 to 113 it is easy to end up with a lastbrowser path still pointing to the old (in ~root/Media/EBooks) directory.  This is now detected, an information alert is shown and the path is reset.
When rotating then opening the file browser the old sized filebrowser would be used.  This is now detected and a new file browser with the same path but appropriate size is used to replace the old one.